#25 isolate fenics/mshr dependencies
