+++ tests/transforms/test_streamline.py
@@ -11,50 +11,6 @@ except ImportError:
 
 @unittest.skipIf('fenics' not in sys.modules,
                  'streamline calculation requires conda-installed fenics')
-class TestCCW(unittest.TestCase):
-
-    def test_ccw_correct(self):
-        """
-            line1 (0, 1) <- (1, 1)
-            line2 (0, 0) -> (1, 0)
-        """
-        line1 = [(1, 1), (0, 1)]
-        line2 = [(0, 0), (1, 0)]
-        result = sl.get_ccw_vertices_from_two_lines(line1, line2)
-        self.assertEqual(result, [(1, 1), (0, 1), (0, 0), (1, 0), (1, 1)])
-
-    def test_ccw_line2_cw(self):
-        """
-            line1 (0, 1) <- (1, 1)
-            line2 (0, 0) <- (1, 0)
-        """
-        line1 = [(1, 1), (0, 1)]
-        line2 = [(1, 0), (0, 0)]
-        result = sl.get_ccw_vertices_from_two_lines(line1, line2)
-        self.assertEqual(result, [(1, 1), (0, 1), (0, 0), (1, 0), (1, 1)])
-
-    def test_ccw_line1_cw(self):
-        """
-            line1 (0, 1) -> (1, 1)
-            line2 (0, 0) -> (1, 0)
-        """
-        line1 = [(0, 1), (1, 1)]
-        line2 = [(0, 0), (1, 0)]
-        result = sl.get_ccw_vertices_from_two_lines(line1, line2)
-        self.assertEqual(result, [(0, 1), (0, 0), (1, 0), (1, 1), (0, 1)])
-
-    def test_ccw_both_cw(self):
-        """
-            line1 (0, 1) -> (1, 1)
-            line2 (0, 0) <- (1, 0)
-        """
-        line1 = [(0, 1), (1, 1)]
-        line2 = [(1, 0), (0, 0)]
-        result = sl.get_ccw_vertices_from_two_lines(line1, line2)
-        self.assertEqual(result, [(0, 1), (0, 0), (1, 0), (1, 1), (0, 1)])
-
-@unittest.skipIf('fenics' not in sys.modules,
-                 'streamline calculation requires conda-installed fenics')
 class TestLaplace(unittest.TestCase):
 
     def test_generate_laplace_field_mesh_values(self):
